refactor: migrate codebase from Java to Kotlin
- add repository classes by Kotlin
- delete domain package of Java
