--- src/main/java/com/group/libraryapp/domain/user/UserRepository.java
@@ -1,12 +1,0 @@
-package com.group.libraryapp.domain.user;
-
-import com.group.libraryapp.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByName(String name);
-
-}
