--- src/main/java/com/group/libraryapp/domain/user/loanhistory/UserLoanHistoryRepository.java
@@ -1,10 +1,0 @@
-package com.group.libraryapp.domain.user.loanhistory;
-
-import com.group.libraryapp.domain.UserLoanHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserLoanHistoryRepository extends JpaRepository<UserLoanHistory, Long> {
-
-  UserLoanHistory findByBookNameAndIsReturn(String bookName, boolean isReturn);
-
-}
