+++ src/main/java/com/group/libraryapp/service/user/UserService.java
@@ -1,6 +1,6 @@
 package com.group.libraryapp.service.user;
 
-import com.group.libraryapp.domain.User;
+import com.group.libraryapp.domain.user.User;
 import com.group.libraryapp.domain.user.UserRepository;
 import com.group.libraryapp.dto.user.request.UserCreateRequest;
 import com.group.libraryapp.dto.user.request.UserUpdateRequest;
