+++ src/main/kotlin/com/group/libraryapp/domain/book/BookRepository.kt
@@ -1,0 +1,9 @@
+package com.group.libraryapp.domain.book
+
+import org.springframework.data.jpa.repository.JpaRepository
+import java.util.*
+
+interface BookRepository : JpaRepository<Book, Long> {
+
+    fun findByName(bookName: String): Optional<Book>
+}
