+++ src/main/kotlin/com/group/libraryapp/domain/user/User.kt
@@ -1,5 +1,7 @@
-package com.group.libraryapp.domain
+package com.group.libraryapp.domain.user
 
+import com.group.libraryapp.domain.book.Book
+import com.group.libraryapp.domain.user.loanhistory.UserLoanHistory
 import javax.persistence.CascadeType
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
