+++ src/main/kotlin/com/group/libraryapp/domain/user/UserRepository.kt
@@ -1,0 +1,9 @@
+package com.group.libraryapp.domain.user
+
+import org.springframework.data.jpa.repository.JpaRepository
+import java.util.Optional
+
+interface UserRepository : JpaRepository<User, Long> {
+
+    fun findByName(name: String): Optional<User>
+}
