+++ src/main/kotlin/com/group/libraryapp/domain/user/loanhistory/UserLoanHistory.kt
@@ -1,5 +1,6 @@
-package com.group.libraryapp.domain
+package com.group.libraryapp.domain.user.loanhistory
 
+import com.group.libraryapp.domain.user.User
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
