+++ src/main/kotlin/com/group/libraryapp/domain/user/loanhistory/UserLoanHistoryRepository.kt
@@ -1,0 +1,8 @@
+package com.group.libraryapp.domain.user.loanhistory
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserLoanHistoryRepository : JpaRepository<UserLoanHistory, Long> {
+
+    fun findByBookNameAndIsReturn(bookName: String, isReturn: Boolean): UserLoanHistory?
+}
