+++ src/test/kotlin/com/group/libraryapp/service/UserServiceTest.kt
@@ -1,6 +1,6 @@
 package com.group.libraryapp.service
 
-import com.group.libraryapp.domain.User
+import com.group.libraryapp.domain.user.User
 import com.group.libraryapp.domain.user.UserRepository
 import com.group.libraryapp.dto.user.request.UserCreateRequest
 import com.group.libraryapp.dto.user.request.UserUpdateRequest
